yeash/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
coinmarketcap==4.2.1
constantly==15.1.0
cryptography==2.1.4
dash==0.21.0
dash-core-components==0.19.0
dash-html-components==0.9.0
dash-renderer==0.11.3
dateparser==0.7.0
decorator==4.2.1
Flask==0.12.2
Flask-Compress==1.4.0
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.0
pandas==0.22.0
pandas-datareader==0.6.0
plotly==2.4.1
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-binance==0.6.4
python-dateutil==2.6.1
pytz==2018.3
regex==2018.2.8
requests==2.18.4
requests-cache==0.4.13
requests-file==1.4.3
requests-ftp==0.3.1
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.2
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.4.3
